+++ packages/playground/html/inline/shared.js
@@ -1,0 +1,4 @@
+import './dep3'
+import { log } from './common'
+
+log('shared')