+++ packages/plugin-react/src/babel.d.ts
@@ -1,0 +1,4 @@
+declare module '@babel/plugin-transform-react-jsx'
+declare module '@babel/plugin-transform-react-jsx-self'
+declare module '@babel/plugin-transform-react-jsx-source'
+declare module 'react-refresh/babel.js'