+++ packages/vite/src/node/__tests__/build.spec.ts
@@ -1,4 +1,5 @@
 import { resolveLibFilename } from '../build'
+import { resolve } from 'path'
 
 describe('resolveLibFilename', () => {
   test('custom filename function', () => {
@@ -8,7 +9,7 @@ describe('resolveLibFilename', () => {
         entry: 'mylib.js'
       },
       'es',
-      'mylib'
+      resolve(__dirname, 'packages/name')
     )
 
     expect(filename).toBe('custom-filename-function.es.js')
@@ -16,17 +17,51 @@ describe('resolveLibFilename', () => {
 
   test('custom filename string', () => {
     const filename = resolveLibFilename(
-      { fileName: 'custom-filename', entry: 'mylib.js' },
+      {
+        fileName: 'custom-filename',
+        entry: 'mylib.js'
+      },
       'es',
-      'mylib'
+      resolve(__dirname, 'packages/name')
     )
 
     expect(filename).toBe('custom-filename.es.js')
   })
 
   test('package name as filename', () => {
-    const filename = resolveLibFilename({ entry: 'mylib.js' }, 'es', 'mylib')
+    const filename = resolveLibFilename(
+      {
+        entry: 'mylib.js'
+      },
+      'es',
+      resolve(__dirname, 'packages/name')
+    )
 
     expect(filename).toBe('mylib.es.js')
+  })
+
+  test('custom filename and no package name', () => {
+    const filename = resolveLibFilename(
+      {
+        fileName: 'custom-filename',
+        entry: 'mylib.js'
+      },
+      'es',
+      resolve(__dirname, 'packages/noname')
+    )
+
+    expect(filename).toBe('custom-filename.es.js')
+  })
+
+  test('missing filename', () => {
+    expect(() => {
+      resolveLibFilename(
+        {
+          entry: 'mylib.js'
+        },
+        'es',
+        resolve(__dirname, 'packages/noname')
+      )
+    }).toThrow()
   })
 })